refactor: move go files to a subdirectory except the main.go
