+++ tpl/functions.go
@@ -1,4 +1,4 @@
-package main
+package tpl
 
 import "text/template"
