+++ tpl/logger.go
@@ -1,4 +1,4 @@
-package main
+package tpl
 
 import (
 	"log"
